Use more convenient way to checkout the project sources.
With that approach the project sources are checked out using
projectUrl, branch and credentials from the job configuration.

In case of using this pipeline script as a central pipeline
script, not contained in the project sources, but served from a
central repo, the projectUrl and the branch of the "payload"
repo needed to be provided via job parameters. This way is
now not supported anymore.

Instead the "payload" repo needs to be referenced always directly in
the job configuration. Hence there is no need anymore for having
repository URL and branch as job parameters. The central pipeline
script is now expected to be loaded from a stub pipeline script
located in the "payload" repo. An example for such a stub pipeline
script is contained in this commit (pipelines/stub/Jenkinsfile).

Advantage of that approach: In case the "payload" repo is protected
with credentials we can use directly use the credentialsId which is
provided in the job configuration together with repoUrl and branch.
Before this was not possible since the repoUrl, the branch and the
credentialsId configured in the job was thatone of the central
pipeline script rather than thatpone of the "payload" repo.

Disadvantate of that approach: The central pipeline script can be
disabled by project members by not using the stub file or by changing
this file.