Handover pipeline script itself into mtabuild
Before moving the validation into the steps we had the validation
right at the beginning of the pipeline. 'this' was used for performing
the validation. In order to be as much backward compatible as possible we
need a reference to the script inside mtaBuild in order to be able to
do the validation in the same manner.